Added new read full file function and changed uint_64 to size_t for size functions

+++ src/os/os_linux.c
@@ -128,11 +128,11 @@ internal void os_file_close(Os_File file)
     close(file);
 }
 
-internal uint64_t os_file_read(Os_File file, Rng1_U64 rng, void *out_data)
+internal size_t os_file_read(Os_File file, Rng1_U64 rng, void *out_data)
 {
-    uint64_t total_num_bytes_to_read = rng1_dim_u64(rng);
-    uint64_t total_num_bytes_read = 0;
-    uint64_t total_num_bytes_left_to_read = total_num_bytes_to_read;
+    size_t total_num_bytes_to_read = rng1_dim_u64(rng);
+    size_t total_num_bytes_read = 0;
+    size_t total_num_bytes_left_to_read = total_num_bytes_to_read;
     for(;total_num_bytes_left_to_read > 0;)
     {
         int read_result = pread(
@@ -152,17 +152,17 @@ internal uint64_t os_file_read(Os_File file, Rng1_U64 rng, void *out_data)
     return total_num_bytes_read;
 }
 
-internal uint64_t os_file_write(Os_File file, Rng1_U64 rng, void *data)
+internal size_t os_file_write(Os_File file, Rng1_U64 rng, void *data)
 {
-    uint64_t total_num_bytes_to_write = rng1_dim_u64(rng);
-    uint64_t total_num_bytes_written = 0;
+    size_t total_num_bytes_to_write = rng1_dim_u64(rng);
+    size_t total_num_bytes_written = 0;
     if (file == OS_STDOUT || file == OS_STDIN || file == OS_STDERR)
     {
         total_num_bytes_written = write(file, data, total_num_bytes_to_write);
     }
     else
     {
-        uint64_t total_num_bytes_left_to_write = total_num_bytes_to_write;
+        size_t total_num_bytes_left_to_write = total_num_bytes_to_write;
         for(;total_num_bytes_left_to_write > 0;)
         {
             int write_result = pwrite(
